+++ lib/Engine/BasicClasses/AllBasicClasses/AllBasicClasses.h
@@ -8,3 +8,4 @@
 #include "../Canvas/Canvas.h"
 #include "../Game/HyperPlane/HyperPlane.h"
 #include "../Game/HyperEllipsoid/HyperEllipsoid.h"
+#include "../EventSystem/EventSystem.h"