+++ lib/Engine/BasicClasses/Game/HyperEllipsoid/HyperEllipsoid.h
@@ -8,7 +8,7 @@ namespace Engine
     class HyperEllipsoid : public Game::Object
     {
     public:
-        HyperEllipsoid(Game::Object object, Point position, Vector diriction, std::vector<float> semiAxes);
+        HyperEllipsoid(Game::Object object, Point position, Vector diriction, Vector semiAxes);
 
         HyperEllipsoid(Entity entity);
 
@@ -17,6 +17,8 @@ namespace Engine
         void Rotate3D(float angle1, float angle2, float angle3);
 
         float IntersectionDistance(Ray ray) override;
+
+        Vector FindParams(Ray ray);
 
         void operator = (Entity entity);