+++ lib/Math/Accuracy/Accuracy.h
@@ -5,4 +5,6 @@ const float PRECISION = 1e-5f;
 
 const int DEC = 1 / PRECISION;
 
+const float INF = 999999999.0f;
+
 float Round(float number);